squash

+++ shelf.go
@@ -155,7 +155,7 @@ func (s *shelf) Close() error {
 	if s.readonly {
 		return nil
 	}
-	defer s.gapCount.Store(len(s.gaps))
+	defer s.gapCount.Store(uint64(len(s.gaps)))
 	defer s.itemCount.Store(s.count)
 	var err error
 	setErr := func(e error) {
@@ -240,7 +240,7 @@ func (s *shelf) Delete(slot uint64) error {
 	if slot >= s.count {
 		return fmt.Errorf("%w: shelf %d, slot %d, tail %d", ErrBadIndex, s.slotSize, slot, s.count)
 	}
-	defer s.gapCount.Store(len(s.gaps))
+	defer s.gapCount.Store(uint64(len(s.gaps)))
 	defer s.itemCount.Store(s.count)
 	// We try to keep writes going to the early parts of the file, to have the
 	// possibility of trimming the file when/if the tail becomes unused.
@@ -314,7 +314,7 @@ func (s *shelf) getSlot() uint64 {
 	// Locate the first free slot
 	s.gapsMu.Lock()
 	defer s.gapsMu.Unlock()
-	defer s.gapCount.Store(len(s.gaps))
+	defer s.gapCount.Store(uint64(len(s.gaps)))
 	defer s.itemCount.Store(s.count)
 	if nGaps := len(s.gaps); nGaps > 0 {
 		slot = s.gaps[0]
@@ -349,7 +349,7 @@ func (s *shelf) Iterate(onData onShelfDataFn) error {
 		nextGap = uint64(0xffffffffffffffff)
 		gapIdx  = 0
 	)
-	defer s.gapCount.Store(len(s.gaps))
+	defer s.gapCount.Store(uint64(len(s.gaps)))
 	defer s.itemCount.Store(s.count)
 	if gapIdx < len(s.gaps) {
 		nextGap = s.gaps[gapIdx]
@@ -381,7 +381,7 @@ func (s *shelf) compact(onData onShelfDataFn) error {
 	defer s.gapsMu.Unlock()
 	s.fileMu.RLock()
 	defer s.fileMu.RUnlock()
-	defer s.gapCount.Store(len(s.gaps))
+	defer s.gapCount.Store(uint64(len(s.gaps)))
 	defer s.itemCount.Store(s.count)
 	buf := make([]byte, s.slotSize)
 	// nextGap searches upwards from the given slot (inclusive),
